+++ components/Button/BlobButton.module.scss
@@ -12,6 +12,7 @@
   line-height: 142%;
 
   width: 100%;
+  white-space: nowrap;
   height: 3.8rem;
   display: inline-flex;
   padding: 0.8rem 1.6rem;
